App now has execute button at top of page
This addition now allows for the user to first input both lat and long, and then press execute to load in the climate data and local map. Before this addition, the app would be really slow and try to load the climate data and local stations map for when the lat was changed, and then the long.