updates
Be sure to re-upload the crops.csv file. Had to remove excess commas so NA's wouldn't take up space in the data frame. Was able to calculate an average crop coefficient based on the selected crops from the user. If we want a weighted average crop coefficient based on the percentage of area for each crop we need to add some kind of input widget for that.